OLMIS-7757: Moved export classes to separate folder
