+++ src/integration-test/java/org/openlmis/referencedata/web/BaseWebIntegrationTest.java
@@ -92,7 +92,6 @@ import org.openlmis.referencedata.repository.UserRepository;
 import org.openlmis.referencedata.repository.custom.impl.ProgramRedisRepository;
 import org.openlmis.referencedata.repository.custom.impl.SupervisoryNodeDtoRedisRepository;
 import org.openlmis.referencedata.service.AuthenticationHelper;
-import org.openlmis.referencedata.service.DataExportService;
 import org.openlmis.referencedata.service.FacilityService;
 import org.openlmis.referencedata.service.FacilityTypeService;
 import org.openlmis.referencedata.service.GeographicZoneService;
@@ -103,6 +102,7 @@ import org.openlmis.referencedata.service.RequisitionGroupService;
 import org.openlmis.referencedata.service.RightService;
 import org.openlmis.referencedata.service.SupplyPartnerBuilder;
 import org.openlmis.referencedata.service.UserService;
+import org.openlmis.referencedata.service.export.DataExportService;
 import org.openlmis.referencedata.testbuilder.UserDataBuilder;
 import org.openlmis.referencedata.util.Message;
 import org.openlmis.referencedata.util.messagekeys.SystemMessageKeys;
