+++ src/main/java/org/openlmis/referencedata/service/TradeItemService.java
@@ -18,6 +18,7 @@ package org.openlmis.referencedata.service;
 import java.util.List;
 import org.openlmis.referencedata.dto.TradeItemCsvModel;
 import org.openlmis.referencedata.repository.TradeItemRepository;
+import org.openlmis.referencedata.service.export.ExportableDataService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
