+++ src/main/java/org/openlmis/referencedata/service/export/ExportableDataService.java
@@ -13,14 +13,14 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
  */
 
-package org.openlmis.referencedata.service;
+package org.openlmis.referencedata.service.export;
 
-import java.io.IOException;
-import java.io.OutputStream;
 import java.util.List;
 
-public interface DataFormatterService {
+public interface ExportableDataService<T> {
 
-  <T> void process(OutputStream outputStream, List<T> data, Class<T> type) throws IOException;
+  List<T> findAllExportableItems();
+
+  Class<T> getExportableType();
 
 }
