+++ src/main/java/org/openlmis/referencedata/web/export/DataExportController.java
@@ -13,13 +13,14 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
  */
 
-package org.openlmis.referencedata.web;
+package org.openlmis.referencedata.web.export;
 
-import static org.openlmis.referencedata.web.DataExportController.RESOURCE_PATH;
+import static org.openlmis.referencedata.web.export.DataExportController.RESOURCE_PATH;
 
 import java.util.Map;
 import org.openlmis.referencedata.domain.RightName;
-import org.openlmis.referencedata.service.DataExportService;
+import org.openlmis.referencedata.service.export.DataExportService;
+import org.openlmis.referencedata.web.BaseController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
