+++ src/main/java/org/openlmis/referencedata/web/export/DataImportController.java
@@ -13,13 +13,14 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
  */
 
-package org.openlmis.referencedata.web;
+package org.openlmis.referencedata.web.export;
 
-import static org.openlmis.referencedata.web.DataImportController.RESOURCE_PATH;
+import static org.openlmis.referencedata.web.export.DataImportController.RESOURCE_PATH;
 
 import java.util.List;
 import org.openlmis.referencedata.dto.BaseDto;
-import org.openlmis.referencedata.service.DataImportService;
+import org.openlmis.referencedata.service.export.DataImportService;
+import org.openlmis.referencedata.web.BaseController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
