+++ src/test/java/org/openlmis/referencedata/service/DataExportServiceTest.java
@@ -24,8 +24,8 @@ import static org.mockito.Mockito.anyString;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.openlmis.referencedata.service.DataExportService.FORMATTER_SERVICE_NAME_SUFFIX;
-import static org.openlmis.referencedata.service.DataExportService.SERVICE_NAME_SUFFIX;
+import static org.openlmis.referencedata.service.export.DataExportService.FORMATTER_SERVICE_NAME_SUFFIX;
+import static org.openlmis.referencedata.service.export.DataExportService.SERVICE_NAME_SUFFIX;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -42,7 +42,11 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.openlmis.referencedata.dto.OrderableDto;
 import org.openlmis.referencedata.exception.ValidationMessageException;
-import org.openlmis.referencedata.web.DataExportParams;
+import org.openlmis.referencedata.service.export.CsvFormatterService;
+import org.openlmis.referencedata.service.export.DataExportService;
+import org.openlmis.referencedata.service.export.DataFormatterService;
+import org.openlmis.referencedata.service.export.ExportableDataService;
+import org.openlmis.referencedata.web.export.DataExportParams;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.core.io.Resource;
